// frontend/src/enums/notification-type.enum.ts
export enum NotificationType {
  // SYSTEM NOTIFICATIONS
  CHECK_IN_REMINDER = 1,
  CHECK_OUT_REMINDER = 2,
  // DEPARTMENT NOTIFICATIONS
  INVITATION_TO_DEPARTMENT = 3,
  UPDATE_DEPARTMENT_INFORMATION = 4,
  DELETE_DEPARTMENT = 5,
  DELETE_OUT_OF_DEPARTMENT = 6,
  // USER PERMISSION NOTIFICATIONS
  NEW_PERMISSION = 7,
  CHANGE_USER_PERMISSION = 8,
  // DEPARTMENT PERMISSION NOTIFICATIONS
  NEW_PERMISSION_IN_DEPARTMENT = 9,
  CHANGE_PERMISSION_IN_DEPARTMENT = 10,
  // LEAVE REQUEST NOTIFICATIONS
  LEAVE_REQUEST = 11,
  APPROVE_LEAVE_REQUEST = 12,
  REJECT_LEAVE_REQUEST = 13,
  CANCEL_LEAVE_REQUEST = 14,
}